dwpal: fieldValuesGet: remove all ' ' characters
The fieldValuesGet function is used by DWPAL to parse lists of fields of
the format "btm_supported=1", SupportedRates-2 4 11 ", "SSID=", etc.

As part of parsing, the implementation adds ' ' characters to the buffer
for handling special cases, and removes them at the end.

The for loop removing the ' ' characters has a bug since it doesn't take
into account the first ' ', so if an empty value is given such as
"SSID=", the returned value will be ' '.

So, fix it by adding a patch for it in dwpal.

Signed-off-by: Tomer Eliyahu <tomer.b.eliyahu@intel.com>