swpal: Fix swpal package Makefile
The current package Makefile won't track changes if after first build we
change the PKG_REV - this is because the commit hash is not part of the
package tarball. This means that when we clean swpal, change the hash in
the Makefile, then prepare it, the DL stage will be skipped since the
tarball already exists in the DL folder, resulting with an old SWPAL.

Fix this by adding the $(PKG_REV) to both the build folder and the
tarball name.

Signed-off-by: Tomer Eliyahu <tomer.b.eliyahu@intel.com>