iwlwav-base-files: Add uci-defaults for unique wlan macs
In UGW, there are SLs (Service Layers) which implement some platform
glue functionality - one of the WLAN SL's roles is to override the
default /etc/config/wireless mac addresses based on the uboot parameter
macaddr which is guaranteed to be unique for the device.

In prplWrt - there are no SLs, therefore the MAC addresses are never
overriden. This results with ALL RAX40 devices using the same WLAN
interfaces MAC addresses - which is BAD.

Fix this by adding a uci-defaults script which reads the base mac
address from /proc/cmdline, and assigns incrementing mac addresses for
the WLAN interfaces defined in /etc/config/wireless which is also part
of the iwlwav-base-files repository - note that this script is tailored
for that configuration file.

The first WLAN interface address is the base mac incremented by 6, in
order to skip the wired interfaces: eth0_0,eth0_1,eth0_2,eth0_3,eth1.

Reference file is the wireless_def_config which is unused in this
repository, so delete it while we're at it.

Signed-off-by: Tomer Eliyahu <tomer.b.eliyahu@intel.com>